Circulations

+++ src/main/java/com/anam/elibrary/controller/BookController.java
@@ -10,6 +10,7 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 
+import javax.servlet.http.HttpSession;
 import java.util.List;
 
 @Controller
@@ -19,7 +20,10 @@ public class BookController {
     private BookService bookService;
 
     @GetMapping("/books")
-    public String books(Model model) {
+    public String books(Model model, HttpSession session) {
+        if (session.getAttribute("username") == null) {
+            return "redirect:/login";
+        }
         List<Book> books = bookService.findAll();
         model.addAttribute("titlePage", "BOOKS");
         model.addAttribute("books", books);
@@ -27,20 +31,29 @@ public class BookController {
     }
 
     @GetMapping("/books/create")
-    public String booksCreate(Model model) {
+    public String booksCreate(Model model, HttpSession session) {
+        if (session.getAttribute("username") == null) {
+            return "redirect:/login";
+        }
         model.addAttribute("titlePage", "BOOKS CREATE");
         model.addAttribute("book", new Book());
         return "dashboard/books-create";
     }
 
     @PostMapping("/books/create")
-    public String booksCreateSubmit(@ModelAttribute(value = "book") Book book) {
+    public String booksCreateSubmit(HttpSession session, @ModelAttribute(value = "book") Book book) {
+        if (session.getAttribute("username") == null) {
+            return "redirect:/login";
+        }
         bookService.save(book);
         return "redirect:/books";
     }
 
     @GetMapping("/books/edit/{id}")
-    public String booksEdit(Model model, @PathVariable("id") int id) {
+    public String booksEdit(Model model, HttpSession session, @PathVariable("id") int id) {
+        if (session.getAttribute("username") == null) {
+            return "redirect:/login";
+        }
         /*Book book = new Book();
         book.setId(12344);
         book.setTitle("APA AJA BOLEH");*/
@@ -56,13 +69,19 @@ public class BookController {
     }
 
     @PostMapping("/books/edit")
-    public String booksEditSubmit(@ModelAttribute(value = "book") Book book) {
+    public String booksEditSubmit(HttpSession session, @ModelAttribute(value = "book") Book book) {
+        if (session.getAttribute("username") == null) {
+            return "redirect:/login";
+        }
         bookService.update(book);
         return "redirect:/books";
     }
 
     @GetMapping("/books/delete/{id}")
-    public String booksDeleteSubmit(@PathVariable("id") int id) {
+    public String booksDeleteSubmit(HttpSession session, @PathVariable("id") int id) {
+        if (session.getAttribute("username") == null) {
+            return "redirect:/login";
+        }
         bookService.deleteById(id);
         return "redirect:/books";
     }
